Move test into existing catch touch test scene
