+++ osu.Game.Rulesets.Catch.Tests/TestSceneCatchTouchInput.cs
@@ -1,10 +1,15 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
+using System.Linq;
 using NUnit.Framework;
 using osu.Framework.Graphics;
 using osu.Framework.Testing;
+using osu.Game.Rulesets.Catch.Beatmaps;
+using osu.Game.Rulesets.Catch.Mods;
 using osu.Game.Rulesets.Catch.UI;
+using osu.Game.Rulesets.Mods;
 using osu.Game.Tests.Visual;
 
 namespace osu.Game.Rulesets.Catch.Tests
@@ -12,11 +17,11 @@ namespace osu.Game.Rulesets.Catch.Tests
     [TestFixture]
     public class TestSceneCatchTouchInput : OsuTestScene
     {
-        private CatchTouchInputMapper catchTouchInputMapper = null!;
-
-        [SetUpSteps]
-        public void SetUpSteps()
+        [Test]
+        public void TestBasic()
         {
+            CatchTouchInputMapper catchTouchInputMapper = null!;
+
             AddStep("create input overlay", () =>
             {
                 Child = new CatchInputManager(new CatchRuleset().RulesetInfo)
@@ -32,12 +37,30 @@ namespace osu.Game.Rulesets.Catch.Tests
                     }
                 };
             });
+
+            AddStep("show overlay", () => catchTouchInputMapper.Show());
         }
 
         [Test]
-        public void TestBasic()
+        public void TestWithoutRelax()
         {
-            AddStep("show overlay", () => catchTouchInputMapper.Show());
+            AddStep("create drawable ruleset without relax mod", () =>
+            {
+                Child = new DrawableCatchRuleset(new CatchRuleset(), new CatchBeatmap(), new List<Mod>());
+            });
+            AddUntilStep("wait for load", () => Child.IsLoaded);
+            AddAssert("check touch input is shown", () => this.ChildrenOfType<CatchTouchInputMapper>().Any());
+        }
+
+        [Test]
+        public void TestWithRelax()
+        {
+            AddStep("create drawable ruleset with relax mod", () =>
+            {
+                Child = new DrawableCatchRuleset(new CatchRuleset(), new CatchBeatmap(), new List<Mod> { new CatchModRelax() });
+            });
+            AddUntilStep("wait for load", () => Child.IsLoaded);
+            AddAssert("check touch input is not shown", () => !this.ChildrenOfType<CatchTouchInputMapper>().Any());
         }
     }
 }
